[VAS-1167] feat: Updated checks on float value

+++ src/main/java/it/gov/pagopa/receipt/pdf/helpdesk/utils/BizEventToReceiptUtils.java
@@ -533,7 +533,7 @@ public class BizEventToReceiptUtils {
             return bizEvent.getTransactionDetails() != null &&
                     new BigDecimal(bizEvent.getPaymentInfo().getAmount()).subtract(
                                     formatEuroCentAmount(bizEvent.getTransactionDetails().getTransaction().getAmount()))
-                            .intValue() == 0;
+                            .floatValue() == 0;
         }
         return true;
     }
